+++ Changelog.md
@@ -2,6 +2,12 @@
 
 ## 0.0.2
 
-* more documentation and examples
-* matrix: implementation of trait Clone
-* vectors: multiply with scalar, add with scalar
+ * more documentation and examples
+ * module `matrix`
+   * implementation of trait Clone for `Matrix`
+   * multiply `Matrix` with a scalar
+   * add a scalar to a `Matrix`
+ * module `vectors`
+   * multiply vector with a scalar
+   * add a scalar to a vector
+ * module `gaussian` to estimate the parameters of a gaussian distribution from a set of samples